Renamed project to "tutorial" so the tabs in MPLAB "Projects" aren't all identical
Note that I also renamed my project folder from dsPIC33CH to dsPIC33CH_tutorial but this broke compilation due to "missing slave.h". The reason is that the generated file: nbproject/Makefile-default.mk contains the absolute path to the slave project, and needs to be deleted in order to force MPLAB to "regenerate" it. This probably goes for the other Makefile-<configuration>.mk files too.